local

+++ NeuralFingerPrint/graph_degree_conv.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-
+import pdb
 
 class GraphDegreeConv(nn.Module):
     def __init__(self, node_size, edge_size, output_size, degree_list, device, batch_normalize=True):
@@ -27,7 +27,7 @@ class GraphDegreeConv(nn.Module):
             node_neighbor_list = neighbor_by_degree[degree]['node']
             edge_neighbor_list = neighbor_by_degree[degree]['edge']
             if degree == 0 and node_neighbor_list:
-                zero = torch.zeros(len(node_neighbor_list), self.output_size).to(self.device).double()
+                zero = torch.zeros(len(node_neighbor_list), self.output_size).to(self.device).float()
                 degree_activation_list.append(zero)
             else:
                 if node_neighbor_list:
@@ -36,14 +36,14 @@ class GraphDegreeConv(nn.Module):
                     # (#nodes, #degree, edge_size)
                     edge_neighbor_repr = edge_repr[edge_neighbor_list, ...]
                     # (#nodes, #degree, node_size + edge_size)
-                    stacked = torch.cat([node_neighbor_repr, edge_neighbor_repr], dim=2)
+                    stacked = torch.cat([node_neighbor_repr.float(), edge_neighbor_repr.float()], dim=2)
                     summed = torch.sum(stacked, dim=1, keepdim=False)
-                    degree_activation = degree_layer(summed)
-                    degree_activation_list.append(degree_activation)
+                    degree_activation = degree_layer(summed.float())
+                    degree_activation_list.append(degree_activation.float())
         neighbor_repr = torch.cat(degree_activation_list, dim=0)
-        self_repr = self.linear(node_repr)
+        self_repr = self.linear(node_repr.float())
         # size = (#nodes, #output_size)
         activations = self_repr + neighbor_repr + self.bias.expand_as(self_repr)
         if self.batch_normalize:
-            activations = self.normalize(activations)
+            activations = self.normalize(activations.float())
         return F.relu(activations)
